remove stubbed unbound modules, move rest of DOM classes to DOM.js
