+++ src/DOM.js
@@ -10,6 +10,7 @@ const util = require('util');
 const {defaultCanvasSize} = require('./constants');
 const {Event, EventTarget, MouseEvent} = require('./Event');
 const GlobalContext = require('./GlobalContext');
+const bindings = require('../native-bindings');
 const symbols = require('./symbols');
 const utils = require('./utils');
 const {_elementGetter, _elementSetter} = require('./utils');
@@ -1649,132 +1650,6 @@ class HTMLSourceElement extends HTMLSrcableElement {
 }
 module.exports.HTMLSourceElement = HTMLSourceElement;
 
-class HTMLImageElement extends HTMLSrcableElement {
-  constructor() {
-    if (typeof arguments[0] === 'number') {
-      const [width = 0, height = 0] = arguments;
-      return new HTMLImageElement([
-        {
-          name: 'width',
-          value: width + '',
-        },
-        {
-          name: 'height',
-          value: height + '',
-        },
-      ], '', null);
-    }
-    const [attrs = [], value = '', location = null] = arguments;
-    super('IMG', attrs, value, location);
-
-    this.data = new Uint8Array(0);
-    // this.stack = new Error().stack;
-
-    this.on('attribute', (name, value) => {
-      if (name === 'src') {
-        process.nextTick(() => { // XXX
-          this.dispatchEvent(new Event('load', {target: this}));
-        });
-      }
-    });
-  }
-
-  get width() {
-    return 0;
-  }
-  set width(width) {}
-
-  get height() {
-    return 0;
-  }
-  set height(height) {}
-
-  get naturalWidth() {
-    return this.width;
-  }
-  set naturalWidth(naturalWidth) {}
-
-  get naturalHeight() {
-    return this.height;
-  }
-  set naturalHeight(naturalHeight) {}
-};
-module.exports.HTMLImageElement = HTMLImageElement;
-
-class HTMLAudioElement extends HTMLMediaElement {
-  constructor(attrs = [], value = '', location = null) {
-    super('AUDIO', attrs, value, location);
-
-    this.readyState = HTMLMediaElement.HAVE_NOTHING;
-
-    this.on('attribute', (name, value) => {
-      if (name === 'src') {
-        this.readyState = HTMLMediaElement.HAVE_ENOUGH_DATA;
-
-        process.nextTick(() => { // XXX
-          this.dispatchEvent(new Event('canplay', {target: this}));
-          this.dispatchEvent(new Event('canplaythrough', {target: this}));
-        });
-      }
-    });
-  }
-
-  get oncanplay() {
-    return _elementGetter(this, 'canplay');
-  }
-  set oncanplay(oncanplay) {
-    _elementSetter(this, 'canplay', oncanplay);
-  }
-
-  get oncanplaythrough() {
-    return _elementGetter(this, 'canplaythrough');
-  }
-  set oncanplaythrough(oncanplaythrough) {
-    _elementSetter(this, 'canplaythrough', oncanplaythrough);
-  }
-}
-HTMLAudioElement.HAVE_NOTHING = HTMLMediaElement.HAVE_NOTHING;
-HTMLAudioElement.HAVE_METADATA = HTMLMediaElement.HAVE_METADATA;
-HTMLAudioElement.HAVE_CURRENT_DATA = HTMLMediaElement.HAVE_CURRENT_DATA;
-HTMLAudioElement.HAVE_FUTURE_DATA = HTMLMediaElement.HAVE_FUTURE_DATA;
-HTMLAudioElement.HAVE_ENOUGH_DATA = HTMLMediaElement.HAVE_ENOUGH_DATA;
-module.exports.HTMLAudioElement = HTMLAudioElement;
-
-class HTMLVideoElement extends HTMLMediaElement {
-  constructor(attrs = [], value = '', location = null) {
-    super('VIDEO', attrs, value, location);
-
-    this.readyState = HTMLMediaElement.HAVE_NOTHING;
-    this.data = new Uint8Array(0);
-
-    this.on('attribute', (name, value) => {
-      if (name === 'src') {
-        this.readyState = HTMLMediaElement.HAVE_ENOUGH_DATA;
-
-        process.nextTick(() => { // XXX
-          this.dispatchEvent(new Event('canplay', {target: this}));
-          this.dispatchEvent(new Event('canplaythrough', {target: this}));
-        });
-      }
-    });
-  }
-
-  get width() {
-    return 0;
-  }
-  set width(width) {}
-  get height() {
-    return 0;
-  }
-  set height(height) {}
-}
-HTMLVideoElement.HAVE_NOTHING = HTMLMediaElement.HAVE_NOTHING;
-HTMLVideoElement.HAVE_METADATA = HTMLMediaElement.HAVE_METADATA;
-HTMLVideoElement.HAVE_CURRENT_DATA = HTMLMediaElement.HAVE_CURRENT_DATA;
-HTMLVideoElement.HAVE_FUTURE_DATA = HTMLMediaElement.HAVE_FUTURE_DATA;
-HTMLVideoElement.HAVE_ENOUGH_DATA = HTMLMediaElement.HAVE_ENOUGH_DATA;
-module.exports.HTMLVideoElement = HTMLVideoElement;
-
 class SVGElement {}
 module.exports.SVGElement = SVGElement;
 
@@ -2097,6 +1972,475 @@ class Comment extends CharacterNode {
   }
 }
 module.exports.Comment = Comment;
+
+class HTMLImageElement extends HTMLSrcableElement {
+  constructor(...args) {
+    if (typeof arguments[0] === 'number') {
+      const [width = 0, height = 0] = arguments;
+      return new HTMLImageElement([
+        {
+          name: 'width',
+          value: width + '',
+        },
+        {
+          name: 'height',
+          value: height + '',
+        },
+      ], '', null);
+    }
+    const [attrs = [], value = '', location = null] = arguments;
+    super('IMG', attrs, value, location);
+
+    this.image = new bindings.nativeImage();
+
+    this.on('attribute', (name, value) => {
+      if (name === 'src') {
+        const src = value;
+
+        const resource = this.ownerDocument.resources.addResource();
+
+        this.ownerDocument.defaultView.fetch(src)
+          .then(res => {
+            if (res.status >= 200 && res.status < 300) {
+              return res.arrayBuffer();
+            } else {
+              return Promise.reject(new Error(`img src got invalid status code (url: ${JSON.stringify(src)}, code: ${res.status})`));
+            }
+          })
+          .then(arrayBuffer => new Promise((accept, reject) => {
+            this.image.load(arrayBuffer, err => {
+              if (!err) {
+                accept();
+              } else {
+                reject(new Error(`failed to decode image: ${err.message} (url: ${JSON.stringify(src)}, size: ${arrayBuffer.byteLength}, message: ${err})`));
+              }
+            });
+          }))
+          .then(() => {
+            this.dispatchEvent(new Event('load', {target: this}));
+          })
+          .catch(err => {
+            console.warn('failed to load image:', src);
+            this.dispatchEvent(new Event('error', {target: this}));
+          })
+          .finally(() => {
+            resource.setProgress(1);
+          });
+      }
+    });
+  }
+
+  get src() {
+    return this.getAttribute('src');
+  }
+  set src(src) {
+    this.setAttribute('src', src);
+  }
+
+  get onload() {
+    return _elementGetter(this, 'load');
+  }
+  set onload(onload) {
+    _elementSetter(this, 'load', onload);
+  }
+
+  get onerror() {
+    return _elementGetter(this, 'error');
+  }
+  set onerror(onerror) {
+    _elementSetter(this, 'error', onerror);
+  }
+
+  get width() {
+    return this.image.width;
+  }
+  set width(width) {}
+  get height() {
+    return this.image.height;
+  }
+  set height(height) {}
+
+  get naturalWidth() {
+    return this.width;
+  }
+  set naturalWidth(naturalWidth) {}
+  get naturalHeight() {
+    return this.height;
+  }
+  set naturalHeight(naturalHeight) {}
+
+  getBoundingClientRect() {
+    return new DOMRect(0, 0, this.width, this.height);
+  }
+
+  get data() {
+    return this.image.data;
+  }
+  set data(data) {}
+};
+module.exports.HTMLImageElement = HTMLImageElement;
+
+class HTMLAudioElement extends HTMLMediaElement {
+  constructor(attrs = [], value = '') {
+    super('AUDIO', attrs, value);
+
+    this.readyState = HTMLMediaElement.HAVE_NOTHING;
+    this.audio = new bindings.nativeAudio.Audio();
+
+    this.on('attribute', (name, value) => {
+      if (name === 'src') {
+        const src = value;
+
+        const resource = this.ownerDocument.resources.addResource();
+
+        this.ownerDocument.defaultView.fetch(src)
+          .then(res => {
+            if (res.status >= 200 && res.status < 300) {
+              return res.arrayBuffer();
+            } else {
+              return Promise.reject(new Error(`audio src got invalid status code (url: ${JSON.stringify(src)}, code: ${res.status})`));
+            }
+          })
+          .then(arrayBuffer => {
+            try {
+              this.audio.load(arrayBuffer);
+            } catch(err) {
+              throw new Error(`failed to decode audio: ${err.message} (url: ${JSON.stringify(src)}, size: ${arrayBuffer.byteLength})`);
+            }
+          })
+          .then(() => {
+            this.readyState = HTMLMediaElement.HAVE_ENOUGH_DATA;
+            this._emit('canplay');
+            this._emit('canplaythrough');
+          })
+          .catch(err => {
+            console.warn('failed to load audio:', src);
+            this.dispatchEvent(new Event('error', {target: this}));
+          })
+          .finally(() => {
+            resource.setProgress(1);
+          });
+      }
+    });
+  }
+
+  play() {
+    this.audio.play();
+  }
+  pause() {
+    this.audio.pause();
+  }
+
+  get currentTime() {
+    return this.audio && this.audio.currentTime;
+  }
+  set currentTime(currentTime) {
+    if (this.audio) {
+      this.audio.currentTime = currentTime;
+    }
+  }
+
+  get duration() {
+    return this.audio && this.audio.duration;
+  }
+  set duration(duration) {
+    if (this.audio) {
+      this.audio.duration = duration;
+    }
+  }
+
+  get oncanplay() {
+    return _elementGetter(this, 'canplay');
+  }
+  set oncanplay(oncanplay) {
+    _elementSetter(this, 'canplay', oncanplay);
+  }
+
+  get oncanplaythrough() {
+    return _elementGetter(this, 'canplaythrough');
+  }
+  set oncanplaythrough(oncanplaythrough) {
+    _elementSetter(this, 'canplaythrough', oncanplaythrough);
+  }
+
+  get onerror() {
+    return _elementGetter(this, 'error');
+  }
+  set onerror(onerror) {
+    _elementSetter(this, 'error', onerror);
+  }
+};
+module.exports.HTMLAudioElement = HTMLAudioElement;
+
+class HTMLVideoElement extends HTMLMediaElement {
+  constructor(attrs = [], value = '', location = null) {
+    super('VIDEO', attrs, value, location);
+
+    this.readyState = HTMLMediaElement.HAVE_NOTHING;
+    this.data = new Uint8Array(0);
+
+    this.on('attribute', (name, value) => {
+      if (name === 'src') {
+        this.readyState = HTMLMediaElement.HAVE_ENOUGH_DATA;
+
+        if (urls.has(value)) {
+          const blob = urls.get(value);
+          if (blob instanceof bindings.nativeVideo.VideoDevice) {
+            this.video = blob;
+          }
+        }
+
+        const resource = this.ownerDocument.resources.addResource();
+
+        process.nextTick(() => {
+          this.dispatchEvent(new Event('canplay', {target: this}));
+          this.dispatchEvent(new Event('canplaythrough', {target: this}));
+
+          resource.setProgress(1);
+        });
+      }
+    });
+  }
+
+  get width() {
+    return this.video ? this.video.width : 0;
+  }
+  set width(width) {}
+  get height() {
+    return this.video ? this.video.height : 0;
+  }
+  set height(height) {}
+
+  get autoplay() {
+    return this.getAttribute('autoplay');
+  }
+  set autoplay(autoplay) {
+    this.setAttribute('autoplay', autoplay);
+  }
+
+  getBoundingClientRect() {
+    return new DOMRect(0, 0, this.width, this.height);
+  }
+
+  get data() {
+    return this.video ? this.video.data : null;
+  }
+  set data(data) {}
+
+  play() {
+    const _getDevice = facingMode => {
+      switch (facingMode) {
+        case 'user': return devices[0];
+        case 'environment': return devices[1];
+        case 'left': return devices[2];
+        case 'right': return devices[3];
+        default: return devices[0];
+      }
+    }
+    const _getName = facingMode => (process.platform === 'darwin' ? '' : 'video=') + _getDevice(facingMode).name;
+    const _getOptions = facingMode => {
+      if (process.platform === 'darwin') {
+        return 'framerate='+_getDevice(facingMode).modes[0].fps;
+      } else {
+        return null;
+      }
+    }
+    if (this.video) {
+      this.video.close();
+      this.video.open(
+        _getName(this.video.constraints.facingMode),
+        _getOptions(this.video.constraints.facingMode)
+      );
+    }
+  }
+  pause() {
+    if (this.video) {
+      this.video.close();
+    }
+  }
+  update() {
+    if (this.video) {
+      this.video.update();
+    }
+  }
+}
+module.exports.HTMLVideoElement = HTMLVideoElement;
+
+/*
+class HTMLVideoElement extends HTMLMediaElement {
+  constructor(attrs = [], value = '') {
+    super('VIDEO', attrs, value);
+
+    this.readyState = HTMLMediaElement.HAVE_NOTHING;
+    this.video = new bindings.nativeVideo.Video();
+
+    this.on('attribute', (name, value) => {
+      if (name === 'src') {
+        console.log('video downloading...');
+        const src = value;
+        this.ownerDocument.defaultView.fetch(src)
+          .then(res => {
+            console.log('video download res');
+            if (res.status >= 200 && res.status < 300) {
+              return res.arrayBuffer();
+            } else {
+              return Promise.reject(new Error(`video src got invalid status code (url: ${JSON.stringify(src)}, code: ${res.status})`));
+            }
+          })
+          .then(arrayBuffer => {
+            console.log('video download arraybuffer');
+            try {
+              this.video.load(arrayBuffer);
+            } catch(err) {
+              throw new Error(`failed to decode video: ${err.message} (url: ${JSON.stringify(src)}, size: ${arrayBuffer.byteLength})`);
+            }
+          })
+          .then(() => {
+            console.log('video download done');
+            this.readyState = HTMLMediaElement.HAVE_ENOUGH_DATA;
+            this._emit('canplay');
+            this._emit('canplaythrough');
+          })
+          .catch(err => {
+            this._emit('error', err);
+          });
+      } else if (name === 'loop') {
+        this.video.loop = !!value || value === '';
+      } else if (name === 'autoplay') {
+        const autoplay = !!value || value === '';
+        if (autoplay) {
+          console.log('video set autoplay');
+          const canplay = () => {
+            console.log('video autoplay play');
+            this.video.play();
+            _cleanup();
+          };
+          const error = () => {
+            _cleanup();
+          };
+          const _cleanup = () => {
+            this.removeListener('canplay', canplay);
+            this.removeListener('error', error);
+          };
+          this.on('canplay', canplay);
+          this.on('error', error);
+        }
+      }
+    });
+  }
+
+  get width() {
+    return this.video.width;
+  }
+  set width(width) {}
+  get height() {
+    return this.video.height;
+  }
+  set height(height) {}
+
+  get loop() {
+    return this.getAttribute('loop');
+  }
+  set loop(loop) {
+    this.setAttribute('loop', loop);
+  }
+
+  get autoplay() {
+    return this.getAttribute('autoplay');
+  }
+  set autoplay(autoplay) {
+    this.setAttribute('autoplay', autoplay);
+  }
+
+  getBoundingClientRect() {
+    return new DOMRect(0, 0, this.width, this.height);
+  }
+
+  get data() {
+    return this.video.data;
+  }
+  set data(data) {}
+
+  play() {
+    this.video.play();
+  }
+  pause() {
+    this.video.pause();
+  }
+
+  get currentTime() {
+    return this.video && this.video.currentTime;
+  }
+  set currentTime(currentTime) {
+    if (this.video) {
+      this.video.currentTime = currentTime;
+    }
+  }
+
+  get duration() {
+    return this.video && this.video.duration;
+  }
+  set duration(duration) {
+    if (this.video) {
+      this.video.duration = duration;
+    }
+  }
+
+  get oncanplay() {
+    return _elementGetter(this, 'canplay');
+  }
+  set oncanplay(oncanplay) {
+    _elementSetter(this, 'canplay', oncanplay);
+  }
+
+  get oncanplaythrough() {
+    return _elementGetter(this, 'canplaythrough');
+  }
+  set oncanplaythrough(oncanplaythrough) {
+    _elementSetter(this, 'canplaythrough', oncanplaythrough);
+  }
+
+  get onerror() {
+    return _elementGetter(this, 'error');
+  }
+  set onerror(onerror) {
+    _elementSetter(this, 'error', onerror);
+  }
+
+  run() {
+    let running = false;
+
+    let sources;
+    const srcAttr = this.attributes.src;
+    if (srcAttr) {
+      this._emit('attribute', 'src', srcAttr.value);
+      running = true;
+    } else if (sources = this.childNodes.filter(childNode => childNode.nodeType === Node.ELEMENT_NODE && childNode.matches('source'))) {
+      for (let i = 0; i < sources.length; i++) {
+        const source = sources[i];
+        const {src} = source;
+        if (src) {
+          this.src = src;
+          running = true;
+          break;
+        }
+      }
+    }
+    const loopAttr = this.attributes.loop;
+    const loopAttrValue = loopAttr && loopAttr.value;
+    if (loopAttrValue || loopAttrValue === '') {
+      this.loop = loopAttrValue;
+    }
+    const autoplayAttr = this.attributes.loop;
+    const autoplayAttrValue = autoplayAttr && autoplayAttr.value;
+    if (autoplayAttrValue || autoplayAttrValue === '') {
+      this.autoplay = autoplayAttrValue;
+    }
+
+    return running;
+  }
+}
+*/
 
 function _hash(s) {
   let result = 0;
